adds create page with confirmation message and delete createThread from home page
